+++ attnroute/__init__.py
@@ -17,7 +17,7 @@ Quick start:
     attnroute status
 """
 
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 __author__ = "jeranaias"
 
 # Core exports